+++ Feelter/Sources/Domain/Entity/Filter/FilterAttributeType.swift
@@ -9,29 +9,29 @@ import Foundation
 
 struct FilterAttribute: Hashable {
     /// 밝기 (-1.0 ~ 1.0)
-    var brightness: Float
+    var brightness: Double
     /// 노출 (-1.0 ~ 1.0)
-    var exposure: Float
+    var exposure: Double
     /// 대비 (-1.0 ~ 1.0)
-    var contrast: Float
+    var contrast: Double
     /// 채도 (0.0 ~ 2.0)
-    var saturation: Float
+    var saturation: Double
     /// 선명도 (-1.0 ~ 1.0) / 셰이더 스케일링 abs(sharpness) * 2.0
-    var sharpness: Float
+    var sharpness: Double
     /// 블러 (-1.0 ~ 1.0) / 셰이더 스케일링 abs(blur) * 10.0
-    var blur: Float
+    var blur: Double
     /// 비네팅 (-1.0 ~ 1.0) / 셰이더 스케일링 vignette * 2.0
-    var vignette: Float
+    var vignette: Double
     /// 노이즈 감소 (-1.0 ~ 1.0) / 셰이더 스케일링 noiseReduction * 0.02
-    var noiseReduction: Float
+    var noiseReduction: Double
     /// 하이라이트 (-1.0 ~ 1.0) / 셰이더 스케일링 highlights + 1.0
-    var highlights: Float
+    var highlights: Double
     /// 그림자 (-1.0 ~ 1.0)
-    var shadows: Float
+    var shadows: Double
     /// 온도 (2000 ~ 10000) / 셰이더 스케일링 (temperature - 6500) / 4000
-    var temperature: Float
+    var temperature: Double
     /// 블랙 포인트 (-1.0 ~ 1.0)
-    var blackPoint: Float
+    var blackPoint: Double
 }
 
 enum FilterAttributeType {
